add deploy sql

+++ deploy/sql/V5__add_field__weigth_product_table.sql
@@ -1,1 +1,1 @@
-ALTER TABLE `products` ADD COLUMN `weigth` double(10) NULL AFTER `unit`;
+ALTER TABLE `products` ADD COLUMN `weigth` float(10,2) NULL AFTER `unit`;
